commands/track: add a --no-excluded flag
We're listing excluded patterns by default to help people identify
patterns deeper in the tree that may conflict with patterns in shallower
parts of the tree.  However, it's possible that people have been parsing
the output of git lfs track, even though it's a porcelain command, not a
plumbing command.  To help these people keep their programs working, add
support for a --no-excluded flag that omits listing excluded paths.

Note that the previous behavior of listing paths that were excluded from
Git LFS as tracked paths is not supported, because it was both incorrect
and confusing.