lfsapi: use read-only Git configuration object in tests
Portions of this test call into the HTTP code, which eventually tries to
write an lfs.*.access = basic setting into .git/config.  Unfortunately,
the .git/config it tries to use is that of the developer's git-lfs
repository, which leads to a large number of substantially similar and
mostly useless entries.

To keep the developer's repository tidy and the test environment
isolated, use a read-only Git configuration object in the tests.  As a
result, this code path will fail silently.  This is good for tests, and
should not impact real use negatively, because an actual client will
simply fall back to determining this automatically.