commands/command_status.go: require a working copy
'git lfs status' is functional in a non-work tree repository only some
of the time. Particularly, when there is nothing to report (and
therefore no files to open), 'git lfs status' work as expected.

However, when there are files changed in the index, Git LFS tries to
open them to see what their hash was before/after the change, and in
particular, if they are/were/remain Git LFS pointers. This feature is
used to power the "Git <sha-1> -> LFS <sha-256>" output of 'git lfs
status'.

When in a bare repository and in the aforementioned scenario, the above
operation will fail to open a file that does not exist, for the
repository is bare and thus does not contain a working tree.

In order to bring about a more consistent experience when using 'git lfs
status', let's match the behavior upstream and exit immediately with the
message:

    $ git lfs status
    This operation must be run in a work tree.