t: add a test for fetching from alternate repositories
Until recently, gitobj did not understand how to read from alternate
repositories.  This caused us to think objects not in the local
repository were missing, and consequently, to cause the object scanner
to exit early and a deadlock.  Since both of these issue are now fixed,
add a test that we can fetch from a shared repository without hangs or
failures.