git: close blob objects when finished
When we read a tree, commit, or tag from gitobj, the object is
automatically closed for us, since these objects are relatively small
and can be read and parsed all in memory.  However, blob objects may be
large, so we return a reader which can be used to read the data from the
object.

In order to properly clean up after ourselves, including closing any
object files which may open, we need to call Close on the blob after
reading it.  Otherwise, we'll eventually run out of file descriptors,
causing us to silently hang.  Restore the reset function that was
removed in e3fcde74 ("git: replace object scanner with one based on
gitobj", 2018-09-05) and ensure we close objects and reset the object
scanner between every call to Scan.