locking: don't modify permissions of ignored files
When a lockable pattern in a gitattributes file matched an ignored file,
we would set the permissions of that file on checkout, despite the fact
that the file was ignored.  This could lead to problems for the user.
Ensure that we only traverse objects tracked by Git when modifying the
permissions of files.