config: allow fetching repo permissions for directories
We currently have a function which provides the appropriate value for
repository permissions for files.  Extend it to handle directories by
copying the read bits to the executable bits, as Git does.

Provide the actual implementation in the tools package, as we will use
it in other packages in the future.