fs: expose repository permissions
We have several places in the code base where we have a Filesystem
object but no Configuration object, namely in the transfer queue.  Pass
the repository permissions down into the Filesystem object and expose
them using the same interface as the config object, so we can use either
one as the instance when calling tools.MkdirAll.