lfs: handle large file sizes on 32-bit systems
When parsing pointers, we parsed the size of the object using ParseInt
with a third argument of 0. This caused the argument to be parsed as an
int, which on 32-bit systems is 32 bits in size. Consequently, when we
had an object larger than 4 GiB in size on a 32-bit system, we would
reject the pointer file as invalid and never realize the object needed
to be pushed, leading to the server side rejecting our pushes due to
missing objects.

Set the size of the integer we're parsing to 64 bits to ensure that we
can always parse a pointer correctly.