lfshttp: close body on redirect
If a redirect occurs, we don't close the HTTP body, forcing us to leave
the current connection open and causing us to eventually run out of file
handles. Ensure we close the body if we're not using it so that we can
reuse the connection, which is both more efficient and nicer to the
remote server.