command/status: handle deleted files gracefully
If a had a file (either Git LFS or plain Git) that was deleted with git
rm but not yet committed, git lfs status would produce an error
indicating that the file could not be opened. Since this is not helpful,
if the user has deleted the file, instead indicate this by showing that
the behavior to be committed is a file deletion.

Note that we will only traverse this code path when the destination Git
hash is all zeros (indicating a deleted or missing file).