docs/man: remove outdated notice about filepath.Match()
Since [1], we have remarked that the `--include` and `--exclude` options
belonging to `git-lfs-fetch(1)` were subject to the file matching rules
of package `path/filepath`'s `Match()` implementation.

Since [2], however, this is no longer accurate: the implementation of
`filepathfilter.New()` converts its arguments to wildmatch-style
semantics, which do _not_ follow the rules of the aforementioned
`path/filepath.Match()`, and instead adhere to Git's guidelines.

The relevant call-chain is as follows:

  1. `commands/command_fetch.go` calls `buildFilepathFilter`.
  2. `buildFilepathFilter` calls `determineIncludeExcludePaths`, and
     then calls `filepathfilter.New()`.
  3. Finally, `filepathfilter.New()` calls `wildmatch.New()` after doing
     some sanitization.

Since the behavior that we are deviating from is adherent to Git's
behavior in similar circumstances, let's avoid explicitly documenting it
as such, since it is a reasonable inference from the context.

[1]: 7926b505 (update fetch include/exclude docs for pattern matching,
     2016-08-16)
[2]: 629a935b (filepathfilter: implement in terms of wildmatch,
     2018-02-07)