docs/man: fix inconsistency in 'git-lfs-ls-files(1)'
In [1], `git lfs ls-files` learned how to print either a '*' or '-' to
indicate a full, or un-smudged LFS pointer, respectively. Likewise, in
[2], the 'git-lfs-ls-files(1)' manual page was updated to include
documentation describing what the new '*' and '-' symbols mean.

Unfortunately, the change in [2] referred to the symbols by their
opposite meaning. Reverse the documentation to make sure that the
meanings listed match the implementation in code.

[1]: 2082b26c (tweak ls-files output, 2015-09-25)
[2]: d7e9960a (Added documentation git-lfs-ls-files' */- output.,
     2017-11-07)