Makefile: ensure trimpath works properly
We recently added code to trim paths out of the binary to aid in
reproducible builds. However, our code had a bug: since there are
multiple sets of double quotes in the build invocation, if the home
directory contained spaces, as it often does on Windows, the build would
fail because the extra unquoted spaces confused the compiler.

To address this issue, remove the extra double quotes.