commands/unlock: always attempt an unlock with --force
When a user has gotten their locks into a bad state, it should be
possible to unlock them by using the ID and --force, even if Git LFS
thinks that the file does not exist or that the repository is for some
reason not in a sane state. Consequently, skip the check for existence
in --force mode and just attempt the operation.

This allows users to recover from an issue where we could produce
relative paths with ".." in them when using junctions on Windows, which
prevented the locks from being unlocked.