command/install: don't warn about lack of root on Windows
On Windows, we don't have a good way to detect administrator-equivalent
permissions. The os.Geteuid call we use on Unix returns -1, so we always
print a warning to the user that they lack root permissions (which,
since it's Windows, which lacks the root user, technically they do).
Nevertheless, this warning is irksome and unwelcome since it's prone to
false positives and doesn't tell the user anything about whether their
operation is actually likely to succeed or fail. Simply skip printing it
on Windows or in any other case where we're unable to fetch the actual
user ID.