docs/api: make lock creation example less confusing
Our locking API docs say that the path for a lock should be relative to
the repository root, but we then use an absolute path in the JSON
example, which is confusing. Improve this by using the same relative
path for the example response as we do for the example request, so that
the example is less confusing.