t: make test server check for required headers
When we make a request to the test server, we need an appropriate Accept
header, and when we make a POST request, we need an appropriate
Content-Type header as well. To ensure we have the right behavior in all
code paths, have the test server check for the Accept header and return
a 406 if it's not correct. Similarly, check all POST requests for an
appropriate Content-Type header, and fail the request if it's wrong.
This ensures that all of our code paths have the expected behavior.