Canonicalize common directory paths on Cygwin
When we look up the Git and root paths together, we canonicalize the
paths on Cygwin to Windows paths. This is required because Cygwin is
essentially a libc, and Go neither uses libc nor has support for Cygwin
natively. Without this, we'd treat absolute Unix-style paths created by
Cygwin's Git as Windows paths relative to the root of the current drive,
breaking most uses of Cygwin-created repositories.

However, we do not canonicalize the paths in several other cases where
we read the output from Git, such as the common directory, which is
used when reading Git objects. Consequently, functionality like git lfs
fsck failed on Cygwin.

To ensure things work as expected on Cygwin, canonicalize the paths in
all cases where we read a directory from Git. In addition, add some
error checking for several of these places as well.