CONTRIBUTING: add a section on issues
Add a section on issues that addresses a few common problems we see when
people report an issue. Encourage users to provide helpful debugging
information like the output of "git lfs env" and GIT_TRACE output, and
to use the troubleshooting section of the wiki to help pin down any
problems. Point out that opening a new issue is preferable to commenting
on a closed issue.

Mention that problems with GitHub are best addressed directly with
GitHub; these are things we aren't likely to be able address here, and
folks reporting them to GitHub are likely to get a better response and
quicker action.