Add support for SOCKS proxies
Currently, we prepend "http://" to the proxy URL if it's not an HTTP or
HTTPS URL. Unfortunately, that breaks support for SOCKS 5 proxies, which
use URLs that start with "socks5://". Fix this by allowing URL schemes
starting with "socks" in addition to those starting with "http".

Note that this does not introduce any support for socks5h proxies, since
Go does not support them, but we will support them automatically once Go
does.