Accept legacy Git SSH URLs
For a long time, the preferred way to write an SSH URL for use with Git
has been with the "ssh" scheme. However, Git still accepts the "git+ssh"
and "ssh+git" schemes for remotes as well, despite marking them as
deprecated in the code. For compatibility with Git, accept these schemes
as an alias for the "ssh" scheme.