README.md: move ttaylorr to the alumni section
For some time I have not been involved in day-to-day Git LFS development
as I once have, thanks entirely to the care that the current core team
has given the project. Let's make official what has already been the
de-facto norm and move myself to the alumni section.

Working on Git LFS for the couple of years that I was involved has been
a great honor and privilege. I very much appreciate the folks that I've
worked with for giving me a chance to grow as an engineer. Likewise, I
appreciate all Git LFS's users who have made Git LFS a delight to work
on.

Thank you all.