lfs: use an absolute path for smudging
On Windows, paths are limited to 260 characters with two exceptions:
manifested applications on Windows 10 may use long path, and all systems
may use UNC paths, which must be absolute. Since we are not currently
using a manifest, use an absolute path when smudging to make so that we
can take advantage of UNC paths on Windows.