track: make --filename work with spaces
When using "git lfs track --filename" with a filename with spaces, the
brackets surrounding the character class for representing spaces were
also escaped, leading to the space not being interpreted correctly.
Update the escaping function we use to handle all of our escaping needs,
performing the escaping in the correct order so that we can handle
spaces correctly.