t: make setup code more resilient when parallelized
When we run tests in parallel, we run the setup code each time.  Since
we share some test data between tests, such as credential files, it's
important that we not overwrite them when running tests in parallel.

If we do, it's possible for the credential helper to find a truncated
file with no credentials, thereby causing the test to fail.  Ensure this
doesn't occur by checking to see if the file exists before overwriting
it.

This doesn't erase the race window entirely, but it decreases it
significantly and limits it to the first two tests to run when we're
running in parallel, since the files will already exist when other tests
are invoked.