Makefile: automatically trim paths in Go 1.13
One persistent problem we've had with building binaries is that paths
from the build system get embedded in them.  We'd like to produce
reproducible binaries, so take advantage of Go 1.13's -trimpath argument
that does this for us.  This is better than the old technique because it
works regardless of where our binaries are built on the system and
regardless of where GOPATH is located.