script/upload: add a --finalize option
When the build and upload process is automated, the only thing left for
the maintainer to do will be to download the assets, sign the hashes,
and upload the signed manifest.  To make this easy and mostly automated,
add an option to the upload script, --finalize, that downloads assets,
optionally allows the maintainer to inspect them, and then uploads the
signed manifest to the GitHub release.