Update wildmatch to v1.0.4
Update wildmatch to v1.0.4 to fix an issue with matching non-ASCII
patterns.

Note that the version number is v1.0.4 instead of v1.0.3 because the
master branch of wildmatch contains changes which are valuable but not
currently compatible with Git LFS and v1.0.3 was mistakenly tagged from
these changes.  These changes would break semantic versioning, so the
tag was deleted and v1.0.4 was tagged with only compatible changes.

We can adopt these incompatible changes in the future, but it's better
to fix this bug now since it is causing pain for multiple users rather
than wait to address these changes until the new wildmatch can be
included.