.github: add a release workflow
Currently, our releases are a highly intricate, mostly manual process.
Some work on automating the process has already occurred, but there is
still a lot of work to be done manually.  This means that a release can
eat up several hours of a maintainer's day, which in turn prevents us
from being able to do patch releases as frequently as we'd like.

Automate almost the entire build process by running it on GitHub
Actions.  This process builds the Windows assets, builds the Linux
packages, and prepares and uploads the release assets to GitHub.  All
that remains is for the maintainer to generate and sign the manifest
themselves and upload it, a process which can be easily handled by
running script/upload with a few arguments.