docker: run interactively only if we have a TTY
When building packages in containers, we pass options to docker.
However, when running in a noninteractive environment such as GitHub
Actions, we can't use the -i option, since standard input will not be a
TTY and docker will consequently fail.  Check for a TTY and pass the -i
option only if one is present.