t/cmd: mark lfstest-realpath as being a test tool
It's important for us to flag this binary as a test tool so that Go
doesn't complain about us having multiple instances of the main function
in different files.  Not doing this causes failures when the Debian
package is built in the Docker containers, since the Debian build system
wants to build all packages, including the test packages, at once.