Makefile: improve signtool handling
When we sign Windows binaries currently, we assume that the user has
loaded the certificate into their system credential store.  In the past,
when binaries were built on bespoke VMs, that was a legitimate design,
but when building binaries in an automated fashion, this is no longer
suitable.

Switch the command line arguments to use dashes instead of slashes,
since Git Bash treats options with leading slashes as paths and rewrites
them, breaking our invocation.  Ensure we don't echo the password in the
output; while GitHub actions should filter things for us, it's better
not to rely on that.  Add the -debug option to help us troubleshoot
problems if they occur.