prune: limit concurrency based on file descriptors
When there's a small number of file descriptors, we can exhaust them all
during a prune, causing it to fail.  This is common on macOS, which has
a limit of 256 file descriptors by default.  Use a heuristic to allocate
FDs based on the limit, and on Windows, which does not support
gerlimit(2), use a heuristic depending on the number of CPUs, as before.