script/cibuild: check for code formatting
Sometimes we get contributions from folks who aren't very familiar with
Go.  This is great, but since they aren't familiar with Go, they don't
know to use go fmt or goimports to tidy files.  This leads to files
that are misformatted, which then causes problems when users who do have
these tools are doing development.

Let's help people check formatting automatically by running it as part
of the CI job.  We already check for trailing whitespace, which is
complementary (since it looks at non-Go files as well), and that check
has been effective, so likely this one will be as well.

There is one potential downside: sometimes output changes between
versions of Go.  This is unfortunate, but we'll need to just pin to
whatever version GitHub Actions uses.