Support fetching entire history of specific refs
This adds support for combining the --all option with refs specified as
positional command-line arguments. This will download all objects
referenced by any commit reachable from any of the specified refs.

The purpose of this is to be able to perform full backups, migrations,
or mirroring operations of only a specific set of refs.