Add support for CentOS 8
Build packages for CentOS 8.  Add it to the list of containers we build
for and the list of packages we have downloads available for.

In the gem spec files, use a different tooling on CentOS 8, since the
old method no longer works there.  Build using a more standard gem build
process so that RPM does the abstraction for us now that it's capable of
it.  Stop deleting the root in the install step since we may have
installed things there by that point, and RPM has handled this properly
for many versions.

Since the syntax "%{?el8}" expands to "1" on CentOS 8 and nothing on
earlier versions, add a 0 in front of it any time we use the macro so
that we don't end up with an empty conditional on earlier versions.
"01" is still true and "0" is false, so the logic works out.

It's been verified that the packages still build on CentOS 7 as well as
CentOS 8.