lfsapi: fall back if there are no system Windows credentials
On Windows, we attempt to use the system credentials when authenticating
using NTLM if they exist.  However, in some cases, the system is not
appropriately configured and we get an error, which causes the push
attempt to abort.

If we get such an error, fall back to prompting for credentials like we
did before 2.8.0.