Allow migrate export to handle non-pointer files gracefully
If a file that was being handled by git lfs migrate export contained the
text "git-lfs", but was not a pointer file, we would abort.  This
occurred because the error was not of a type that indicated that the
file was not a pointer, so we treated it as an unrecoverable error
instead of a recoverable one.

Wrap the error in an appropriate type so that we can recover from this
and treat it as not being a pointer.  Add tests for both error branches
in the function to ensure we recover from them.