status: update index before showing status
When running `git lfs status`, we perform a `git diff-index`.  However,
we don't update the index first, so any changes, such as permissions
changes due to locking, cause the file to be listed as modified.  Since
these changes don't represent actual changes that we're interested in,
refresh the index before running diff-index so that it doesn't produce
spurious output.