tq: retry batch failures
If the batch operation fails due to an error instead of a bad HTTP
status code, we'll abort the batch operation and retry.  This appears to
be a regression from 1412d6e4 ("Don't fail if we lack objects the server
has", 2019-04-30), which caused us to handle errors differently.

Since there are two error returns from enqueueAndCollectRetriesFor,
let's wrap the batch error case as a retriable error and not abort if we
find a retriable error later on.  This lets us continue to abort if we
get a missing object, which should be fatal, but retry in the more
common network failure case.