rpm: fix order of exclude arguments
CentOS 8 has recently gained a new version of tar.  This version of tar
is sensitive about the order of its operands and only applies the
--exclude arguments to file arguments following them.  As a result, this
version of tar ignores the exclude arguments altogether, which results
in us trying to recursively copy the contents of the tmptar directory
into itself until the operation fails because the file name is too long.

Predictably, this results in a failed build for CentOS 8.  To avoid
that, place the --exclude options before the file arguments so that
they apply correctly and the build succeeds.