workflows/ci: build Linux packages
Currently we only build Linux packages during the release process.  This
means that there is a significant period of time during which unrelated
changes could silently break the Linux package build.  Since having this
happen is a hassle resulting in the maintainer doing the release having
to build the packages by hand, we should check this process during CI to
ensure that any errors are caught early.

Add a CI job which builds Linux packages as normal but doesn't upload
them anywhere, so that we can ensure that the packages continue to
build.