rpm: use old setup code on CentOS 7
We need to have different %setup invocations on CentOS 8 and older
CentOS versions.  Originally, CentOS 7 was thought to behave like CentOS
8, but the fact that it behaves like CentOS 6 was hidden by the
container images building the RPMs against Git LFS 2.4.0 as part of the
setup process.  As a consequence, we never built the RPMs using the new
versions of the spec files in master, and we never noticed that they
failed in this case.

Since we're no longer building RPMs against an ancient version of Git
LFS as part of the setup process, adjust the Ruby RPMs to call the
correct %setup invocation on CentOS 7.  Note that like the code with
%{?el6}, %{?el7} will expand to either 1 or nothing, so the condition
will be true (nonzero) for CentOS 6 or 7 and false (zero) for 8.