.github/workflows: use actions/checkout v1
When building artifacts during CI, we need the tags in order to be able
to use git describe.  The master version of the checkout action provides
only the one ref we've checked out, but the v1 provides us with a full
clone.  Use the v1 clone so that we can ensure that our CI works as
expected.