README: explain how to verify releases
Some folks aren't familiar with the signed SHA-256 sums file we use, or
aren't sure how to get the keys to verify releases.  To help them, point
them to the tag we've created that contains the keys (similar to the way
Git does this) and add an explanation to the README to help them verify
the releases.