.github/workflows: disable cgo when building on Linux
In the previous commit, we removed the need for building with cgo on
Linux.  However, Go still wants to link against glibc by default,
probably due to support for the name service switch.  This makes our
binaries not work on musl-based systems, however.

Since we already build the RPMs without cgo, it's likely that nobody
actually cares about support for the small number of resolver features
we'd be missing, so explicitly build with cgo when building binaries in
workflows.  Note that any binaries built on amd64 Linux for other
platforms (whether Linux or not) are already not using cgo, since
they're cross-compiled and we wouldn't have the appropriate compilers or
libcs available to support cgo anyway, so this only affects the amd64
binaries we produce.