commands: print help output with --help
If the user invokes "git-lfs --help", we tell them that there's no usage
text found for "--help".  This isn't very helpful.  Since what the user
probably wanted was to invoke the general help output, let's do that by
rewriting the command to "git-lfs", which is the topic for the general
help output.